monitoring/docker/requirements.txt
prometheus-client
PyYAML
paramiko
requests
